refactor(extract-stream): moved out of util
I would rather keep util for extractable/reusable/generic utilities.
extract-stream has domain-specific knowledge.
