+++ lib/extract-stream.js
@@ -1,6 +1,6 @@
 'use strict'
 
-var gunzip = require('./gunzip-maybe')
+var gunzip = require('./util/gunzip-maybe')
 var path = require('path')
 var pipeline = require('mississippi').pipeline
 var tar = require('tar-fs')
